Created screen segments
